+++ lib/responsive_sizer.dart
@@ -1,5 +1,8 @@
 library responsive_sizer;
 
+import 'dart:io' show Platform;
+
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/widgets.dart';
 
 part 'src/extension.dart';